Вынес инициаллизацию контроллеров в функцию
